+++ tests/util/locale.sh
@@ -3,14 +3,10 @@ mkdir -p "util/$BASETESTDIR/locale/$TEST_LANGUAGE/LC_MESSAGES"
 ln -s "../../../../../../i18n/test-${TEST_LANGUAGE}.mo" "util/$BASETESTDIR/locale/$TEST_LANGUAGE/LC_MESSAGES/dtee.mo"
 
 # Find the best possible system language to use when running the test
-FOUND_LOCALES=("" "" "" "" "" "")
+FOUND_LOCALES=("" "" "" "" "")
 while read -r line; do
 	case "$line" in
-		C|POSIX)
-			;;
-
-		C.utf8|C.UTF8|C.utf-8|C.UTF-8)
-			FOUND_LOCALES[5]="$line"
+		C|C.*|POSIX)
 			;;
 
 		"$TEST_LANGUAGE".utf8|"$TEST_LANGUAGE".UTF8|"$TEST_LANGUAGE".utf-8|"$TEST_LANGUAGE".UTF-8)
@@ -40,6 +36,11 @@ for ((i = 0; i < count; i++)); do
 	[ ! -z "${FOUND_LOCALES[i]}" ] && FOUND_LOCALE="${FOUND_LOCALES[i]}" && set -x && return
 done
 
-echo "No locales found. At least one locale other than C or POSIX must be"
-echo "present for gettext() to translate messages."
-exit 1
+echo "No usable locales found. At least one locale other than C, C.* or POSIX must"
+echo "be present for gettext() to translate messages."
+locale -a
+if [ "$CI" = "true" ]; then
+	exit $TEST_EX_FAIL
+else
+	exit $TEST_EX_SKIP
+fi